requirements.txt
# Dépendances principales
pymupdf
sentence-transformers
faiss-cpu
streamlit
openai
langdetect
joblib
numpy
